+++ .github/workflows/ci.yml
@@ -44,10 +44,17 @@ jobs:
           fetch-depth: 1
 
       - name: Install dask-ms
-        run: pipenv install .[testing]
+        run: pipenv install .[testing,zarr]
 
       - name: Run base test suite
         run: pipenv run py.test --flake8 -s -vvv daskms/
+
+      - name: Install dask-ms with xarray support
+        run: pipenv install .[testing,xarray,zarr]
+
+      - name: Run complete test suite
+        run: pipenv run py.test --flake8 -s -vvv daskms/
+
 
   deploy:
     needs: [test]
